pyproject.toml
[project]
name = "fake-job-postings"
version = "0.1.0"
description = ""
authors = [
    {name = "aspaul20",email = "amursaqib1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "pandas (>=2.3.0,<3.0.0)",
    "polars (>=1.31.0,<2.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "lightning (>=2.5.2,<3.0.0)",
    "transformers (>=4.53.0,<5.0.0)",
    "scikit-learn (==1.6)",
    "pytorch-lightning (>=2.5.2,<3.0.0)",
    "comet (>=3.1.0,<4.0.0)",
    "comet-ml (>=3.49.11,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)"
]

[tool.poetry]
packages = [{ include = "fake_job_postings", from = "src" }]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
ruff = "^0.12.1"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0", "setuptools<81"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
train = "src.fake_job_postings.main:train"

[tool.ruff]
src = ["src", "tests"]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
line-ending = "auto"